Move serialized suite actions into SerializedSuiteController
